moved some small components into new files
